Create CityExchangeSup supervisor
Current architecture has a couple of issues - whenever we try generating
new number for nonexistent city, we crash both City, which keeps our
cities with prefixes, and CityExchange processes with phone numbers.
Lets improve that.

We have created CityExchangeSup supervisor. Just like previously, we
specify children and their restart strategy. But this time we use
:simple_one_for_one supervisor.

This strategy is not like the other. Instead of starting processes on
start, we only gain ability to dynamically start children on the fly.
Whenever we need another city holder, we call
Supervisor.start_child(supervisor_module, [arguments passed to child]).

We can also dynamically start children on other types of supervisors,
but this one is more lightweight.